requirements.txt
tkinter
pyaudio
openai-whisper
torch
keyboard
pyautogui
pyperclip
openai